Extract FeedbackService from the controller
Move all bug-report construction — reporter/environment resolution, mail
template assembly, and GitHub issue title/body building — into a new
FeedbackService with buildBugReport / sendBugReportEmail / fileBugReportIssue.
The controller now only checks the permission and calls the service.
