+++ backend/src/api/feedback/feedback.module.ts
@@ -3,9 +3,11 @@ import { GithubModelModule } from "src/models/github/github-model.module";
 import { MailModelModule } from "src/models/mail/mail-model.module";
 import { UsersModelModule } from "src/models/users/users-model.module";
 import { FeedbackController } from "./controllers/feedback.controller";
+import { FeedbackService } from "./services/feedback.service";
 
 @Module({
 	controllers: [FeedbackController],
 	imports: [MailModelModule, UsersModelModule, GithubModelModule],
+	providers: [FeedbackService],
 })
 export class FeedbackModule {}
